// kraken-web/src/gen/kraken_pb.ts
export enum MeshTransportType {
  /**
   * @generated from enum value: MESH_TRANSPORT_UNKNOWN = 0;
   */
  MESH_TRANSPORT_UNKNOWN = 0,

  /**
   * @generated from enum value: MESH_TRANSPORT_SMB = 1;
   */
  MESH_TRANSPORT_SMB = 1,

  /**
   * @generated from enum value: MESH_TRANSPORT_TCP = 2;
   */
  MESH_TRANSPORT_TCP = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(MeshTransportType)
proto3.util.setEnumType(MeshTransportType, "kraken.MeshTransportType", [
  { no: 0, name: "MESH_TRANSPORT_UNKNOWN" },
  { no: 1, name: "MESH_TRANSPORT_SMB" },
  { no: 2, name: "MESH_TRANSPORT_TCP" },
]);

/**
 * @generated from enum kraken.MeshRoleType
 */
export enum MeshRoleType {
  /**
   * @generated from enum value: MESH_ROLE_LEAF = 0;
   */
  MESH_ROLE_LEAF = 0,

  /**
   * @generated from enum value: MESH_ROLE_RELAY = 1;
   */
  MESH_ROLE_RELAY = 1,

  /**
   * @generated from enum value: MESH_ROLE_HUB = 2;
   */
  MESH_ROLE_HUB = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(MeshRoleType)
proto3.util.setEnumType(MeshRoleType, "kraken.MeshRoleType", [
  { no: 0, name: "MESH_ROLE_LEAF" },
  { no: 1, name: "MESH_ROLE_RELAY" },
  { no: 2, name: "MESH_ROLE_HUB" },
]);

/**
 * @generated from enum kraken.MeshLinkState
 */
export enum MeshLinkState {
  /**
   * @generated from enum value: MESH_LINK_CONNECTING = 0;
   */
  MESH_LINK_CONNECTING = 0,

  /**
   * @generated from enum value: MESH_LINK_HANDSHAKING = 1;
   */
  MESH_LINK_HANDSHAKING = 1,

  /**
   * @generated from enum value: MESH_LINK_ACTIVE = 2;
   */
  MESH_LINK_ACTIVE = 2,

  /**
   * @generated from enum value: MESH_LINK_DEGRADED = 3;
   */
  MESH_LINK_DEGRADED = 3,

  /**
   * @generated from enum value: MESH_LINK_FAILED = 4;
   */
  MESH_LINK_FAILED = 4,
}